Final version of zyc's code
This is the final version of my code. but it is without the new found variables. Since Andy uploaded the datasets a bit late, I may not have enough time to adjust my code and find some more interesting things.
As I am afraid to make the main branch and the main file be wrong again, I just write my code locally.